Display details of individual files
Refactor processing of individual files into a process_file function
that is also reused when parsing a directory.

When receiving a path to a specific file now genplis will print its
details: all parsed tags if it's a music file, parsed filters when it's
an M3UG file.

I bumped this up in the roadmap because I was facing a bug where some
files were being included in some playlists when they shouldn't. When
looking at the parsed tags it became clear why: some files had a
"rating" tag that was taking priority over "fmps_rating".